backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
python-dotenv==1.0.0
groq==0.11.0
httpx==0.27.0
redis==5.0.1
celery==5.3.6
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
resend==0.7.0
